# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing database connections and syncing data to destinations like Typesense. List endpoints for sync jobs and sync logs return a paginated envelope: items, page, page_size, total, total_pages.",
        "title": "Portico Sync API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/sync-logs": {
            "get": {
                "description": "Returns a paginated list of sync logs, newest first. Optionally filter by sync_job_id. Default page size is 20; maximum is 100. Each item includes the related sync job when available.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sync-logs"
                ],
                "summary": "List sync logs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by sync job ID",
                        "name": "sync_job_id",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "Page number (1-based)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page (max 100)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SyncLogListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/sync-logs/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sync-logs"
                ],
                "summary": "Get sync log",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sync log ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SyncLog"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/sync-logs/{id}/stop": {
            "post": {
                "description": "Cancels a background sync and sets the log status to stopped.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sync-logs"
                ],
                "summary": "Stop a running sync log",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sync log ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SyncLog"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "handlers.SyncLogListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SyncLog"
                    }
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "page_size": {
                    "type": "integer",
                    "example": 20
                },
                "total": {
                    "type": "integer",
                    "example": 100
                },
                "total_pages": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "models.Connection": {
            "type": "object",
            "properties": {
                "config": {
                    "type": "object"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.SyncJob": {
            "type": "object",
            "properties": {
                "chunk_size": {
                    "type": "integer"
                },
                "config": {
                    "type": "object"
                },
                "created_at": {
                    "type": "string"
                },
                "destination_connection": {
                    "$ref": "#/definitions/models.Connection"
                },
                "destination_connection_id": {
                    "type": "integer"
                },
                "destination_table": {
                    "type": "string"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SyncJobField"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "relations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SyncJobRelation"
                    }
                },
                "rules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SyncJobRule"
                    }
                },
                "source_connection": {
                    "$ref": "#/definitions/models.Connection"
                },
                "source_connection_id": {
                    "type": "integer"
                },
                "source_table": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "workers": {
                    "type": "integer"
                }
            }
        },
        "models.SyncJobField": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "destination_config": {
                    "type": "object"
                },
                "destination_name": {
                    "type": "string"
                },
                "destination_type": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "source_name": {
                    "type": "string"
                },
                "sync_job_id": {
                    "type": "integer"
                },
                "sync_job_relation_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.SyncJobRelation": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "config": {
                    "type": "object"
                },
                "created_at": {
                    "type": "string"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SyncJobField"
                    }
                },
                "foreign_key": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "integer"
                },
                "related_key": {
                    "type": "string"
                },
                "sync_job_id": {
                    "type": "integer"
                },
                "table": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.SyncJobRule": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "field": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "operator": {
                    "type": "string"
                },
                "sync_job_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "models.SyncLog": {
            "type": "object",
            "properties": {
                "duration_ms": {
                    "type": "integer"
                },
                "finished_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "rows_synced": {
                    "type": "integer"
                },
                "rows_total": {
                    "type": "integer"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "sync_job": {
                    "$ref": "#/definitions/models.SyncJob"
                },
                "sync_job_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}